+++ podasync/src/main/java/com/fanap/podasync/Async.java
@@ -257,49 +257,49 @@ public class Async extends WebSocketAdapter {
                         String token, String ssoHost, String deviceID) {
 
         WebSocketFactory webSocketFactory = new WebSocketFactory();
-//        SSLSocketFactory factory = new SSLSocketFactory() {
-//            @Override
-//            public Socket createSocket(String host, int port) throws IOException, UnknownHostException {
-//                return null;
-//            }
+        SSLSocketFactory sslSocketFactory = new SSLSocketFactory() {
+            @Override
+            public String[] getDefaultCipherSuites() {
+                return new String[0];
+            }
+
+            @Override
+            public String[] getSupportedCipherSuites() {
+                return new String[0];
+            }
+
+            @Override
+            public Socket createSocket(Socket s, String host, int port, boolean autoClose) throws IOException {
+                return null;
+            }
+
+            @Override
+            public Socket createSocket(String host, int port) throws IOException, UnknownHostException {
+                return null;
+            }
+
+            @Override
+            public Socket createSocket(String host, int port, InetAddress localHost, int localPort) throws IOException, UnknownHostException {
+                return null;
+            }
+
+            @Override
+            public Socket createSocket(InetAddress host, int port) throws IOException {
+                return null;
+            }
+
+            @Override
+            public Socket createSocket(InetAddress address, int port, InetAddress localAddress, int localPort) throws IOException {
+                return null;
+            }
+        };
+//        webSocketFactory.setSSLSocketFactory()
+//        SSLContext context =
 //
-//            @Override
-//            public Socket createSocket(String host, int port, InetAddress localHost, int localPort) throws IOException, UnknownHostException {
-//                return null;
-//            }
-//
-//            @Override
-//            public Socket createSocket(InetAddress host, int port) throws IOException {
-//                return null;
-//            }
-//
-//            @Override
-//            public Socket createSocket(InetAddress address, int port, InetAddress localAddress, int localPort) throws IOException {
-//                return null;
-//            }
-//
-//            @Override
-//            public String[] getDefaultCipherSuites() {
-//                return new String[0];
-//            }
-//
-//            @Override
-//            public String[] getSupportedCipherSuites() {
-//                return new String[0];
-//            }
-//
-//            @Override
-//            public Socket createSocket(Socket s, String host, int port, boolean autoClose) throws IOException {
-//                return null;
-//            }
-//        };
-//        webSocketFactory.setSSLSocketFactory(factory);
-        // Create a custom SSL context.
-//        SSLContext context = NaiveSSLContext.getInstance("TLS");
-// Set the custom SSL context.
-//        webSocketFactory.setSSLContext(context);
-//        webSocketFactory.setVerifyHostname(false);
-        SSLSocketFactory.getDefault();
+////                NaiveSSLContext.getInstance("TLS");
+////        webSocketFactory.setSSLContext(context);
+////        webSocketFactory.setVerifyHostname(false);
+////        SSLSocketFactory.getDefault();
         saveDeviceId(deviceID);
         setAppId(appId);
         setServerAddress(socketServerAddress);
@@ -486,6 +486,7 @@ public class Async extends WebSocketAdapter {
         if (BuildConfig.DEBUG) Logger.i("READY FOR CHAT", textMessage);
         try {
             asyncListenerManager.callOnStateChanged("CHAT_READY");
+            stateLiveData.postValue("CHAT_READY");
         } catch (IOException e) {
             e.printStackTrace();
         }