+++ podchat/src/main/java/com/fanap/podchat/persistance/AppDatabase.java
@@ -6,9 +6,10 @@ import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
 
 import com.fanap.podchat.mainmodel.Contact;
+import com.fanap.podchat.mainmodel.Thread;
 import com.fanap.podchat.persistance.dao.MessageDao;
 
-@Database(entities = Contact.class, version = 1, exportSchema = false)
+@Database(entities = {Contact.class}, version = 1, exportSchema = false)
 public abstract class AppDatabase extends RoomDatabase {
 
     public static final String DATABASE_DB = "cache.db";